Write-output-file function added
Also parse_output in matlab fun added.

Minor changes in original matlab code as well

.gitignore file updated

+++ matlab/parse_data_bin.m
@@ -3,6 +3,7 @@ clear;
 
 load('r15.mat');
 
+
 fileID = fopen('../data/r15.bin', 'w+');
 
 [N,D] = size(X);
@@ -14,6 +15,6 @@ fwrite(fileID, D, 'int32');
 % Write Main Body [ x11,x12, ... x1D, x21 ... x2D ... xN1 ... xND];
 tempX = X';
 % tempX = [1:(N*D)];
-fwrite(fileID, tempX(:), 'double'); % Data are written in column order...
+fwrite(fileID, tempX(:), 'float'); % Data are written in column order...
 
 fclose(fileID);
